% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domaine.R
\name{calcul_domaine}
\alias{calcul_domaine}
\title{Calculer le domaine de validité d'une fonction}
\usage{
calcul_domaine(f, borne_inf = -100, borne_sup = 100)
}
\arguments{
\item{f}{La fonction dont on veut déterminer le domaine de validité.}

\item{borne_inf}{La borne inférieure de la gamme de valeurs à tester.}

\item{borne_sup}{La borne supérieure de la gamme de valeurs à tester.}
}
\value{
Un vecteur contenant toutes les valeurs de x pour lesquelles f peut être évalué sans erreur.
}
\description{
Cette fonction détermine le domaine de validité d'une fonction pour une gamme de valeurs.
}
\examples{
f <- function(x) sqrt(x)
borne_inf <- -10
borne_sup <- 10
domaine(f, borne_inf, borne_sup)
}
