% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limites.R
\name{calcul_limite}
\alias{calcul_limite}
\title{Calculer la limite d'une fonction}
\usage{
calcul_limite(expression, variable, point)
}
\arguments{
\item{expression}{L'expression de la fonction dont on veut déterminer la limite.}

\item{variable}{La variable de la fonction dont on veut déterminer la limite.}

\item{point}{Le point où on calcule la limite.}
}
\value{
La limite de la fonction.
}
\description{
Cette fonction utilise le package Ryacas pour calculer la limite d'une fonction à un point donné.
}
\examples{
expression <- "x^2 - 3*x + 2"
variable <- "x"
point <- 1
calcul_limite(expression, variable, point)
}
