% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construire_domaine.R
\name{construire_domaine_C1}
\alias{construire_domaine_C1}
\title{'construire_domaine_C1' est une fonction qui construit un domaine où tous les paramètres sont de classe C1.}
\usage{
construire_domaine_C1(f, borne_inf = -10, borne_sup = 10)
}
\arguments{
\item{f}{La fonction dont on veut déterminer le domaine de validité.}

\item{borne_inf}{La borne inférieure de la gamme de valeurs à tester.}

\item{borne_sup}{La borne supérieure de la gamme de valeurs à tester.}
}
\value{
Un vecteur contenant toutes les valeurs de x pour lesquelles f et sa dérivée peuvent être évaluées sans erreur.
}
\description{
'construire_domaine_C1' est une fonction qui construit un domaine où tous les paramètres sont de classe C1.
}
\examples{
f <- function(x){return(sin(x))}
construire_domaine_C1(f)
library(ggplot2)
}
