% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riccati.R
\name{declare_riccati}
\alias{declare_riccati}
\title{'riccati' est une fonction qui represente les parametre de l'quation de riccati.}
\usage{
declare_riccati(a, b, c, y0)
}
\arguments{
\item{a}{La fonction a(x) ecrit sous forme de chaine de caractère}

\item{b}{La fonction b(x) ecrit sous fprme de chaine de caractère}

\item{c}{La fonction c(x) écrit sous forme de chaine de caractère}

\item{y0}{La fonction y0(x) description qui est la solution particulère de l'équation différentielle de bernouilli ecrit sous forme de chaine de c}

\item{y}{La fonction y}

\item{x}{La variable x}
}
\value{
La valeur de l'équation de Riccati.
}
\description{
y' = a + b\emph{y + c}y^2
}
\examples{
a <- function(t) { return(sin(t)-exp(-t^2)) }
b <- function(t) { return(cos(t^3)) }
d <- function(t) { return(tan(t)) }
y_particular <- function(t) { return(exp(t)) }
riccati(a, b, c,y_particular)

}
