% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivee.R
\name{derive}
\alias{derive}
\title{'calculer_derivee' est une fonction qui calcule la dérivée d'une fonction donnée.}
\usage{
derive(expr, var = "x")
}
\arguments{
\item{expr}{L'expression mathématique à dériver}

\item{var}{La variable par rapport à laquelle dériver
@return Un objet représentant la dérivée de l'expression par rapport à la variable.
@example
derive(quote(x^2))
derive(quote(sin(x)))
Calculer la dérivée d'une expression mathématique}
}
\value{
La dérivée de l'expression
}
\description{
Args:
}
