% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_numeric.R
\name{derivee_numerique}
\alias{derivee_numerique}
\title{Calcul de la dérivée numérique d'une fonction}
\usage{
derivee_numerique(f, x, h = 1e-05)
}
\arguments{
\item{f}{La fonction dont on veut calculer la dérivée.}

\item{x}{Le point où la dérivée doit être calculée.}

\item{h}{La valeur de h utilisée dans la formule de la différence centrée.
Par défaut, elle est fixée à .Machine$double.eps^0.25.}
}
\value{
La dérivée de la fonction f au point x.
}
\description{
Cette fonction calcule la dérivée numérique d'une fonction à un point donné.
Elle utilise une approximation de la dérivée basée sur une différence centrée,
(f(x + h) - f(x - h)) / (2 * h), où h est une petite valeur.
}
\examples{
derivee_numerique(function(x) x^2, 2)
}
