% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivabilite.R
\name{est_derivale}
\alias{est_derivale}
\title{Vérifier si une fonction est dérivable}
\usage{
est_derivale(f, h = .Machine$double.xmin)
}
\arguments{
\item{f}{Une fonction à vérifier.}

\item{h}{Un petit nombre utilisé pour le calcul de la différence finie. Par défaut, h = 1e-7.}
}
\value{
\code{TRUE} si la fonction est dérivable, \code{FALSE} sinon.
}
\description{
Cette fonction tente de calculer la dérivée d'une fonction pour vérifier si elle est dérivable.
}
