% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_bernouilli.R
\name{riccati_to_bernoulli}
\alias{riccati_to_bernoulli}
\title{'riccati_to_bernoulli' est une fonction qui transforme une équation de Riccati en une équation de Bernoulli.}
\usage{
riccati_to_bernoulli(a, b, d, y_particular)
}
\arguments{
\item{a}{La fonction a(x)}

\item{b}{La fonction b(x)}

\item{d}{La fonction d(x)}

\item{y_particular}{Une solution particulière de l'équation de Riccati}
}
\value{
Les paramètres de l'équation de Bernoulli associée.
}
\description{
y' = a + b\emph{y + c}y^2
Si y_particular est une solution particulière de l'équation de Riccati,
alors z = 1/(y - y_particular) est une solution de l'équation de Bernoulli
z' + (b - c\emph{y_particular)\emph{z = c
y' = a + b}y + c}y^2
}
\examples{
a <- function(t) { return(sin(t)-exp(-t^2)) }
b <- function(t) { return(cos(t^3)) }
d <- function(t) { return(tan(t)) }
y_particular <- function(t) { return(exp(t)) }
riccati_to_bernoulli(a, b, d, y_particular)
}
