% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue.R
\name{verifier_continuite}
\alias{verifier_continuite}
\title{Vérifier la continuité d'une fonction}
\usage{
verifier_continuite(f, borne_inf = -100, borne_sup = 100, epsilon = 1e-06)
}
\arguments{
\item{f}{La fonction à vérifier.}

\item{borne_inf}{La borne inférieure de la gamme de valeurs à tester.}

\item{borne_sup}{La borne supérieure de la gamme de valeurs à tester.}
}
\value{
Un vecteur logique indiquant si la fonction est continue pour chaque point dans son domaine.
}
\description{
Cette fonction détermine si une fonction est continue sur son domaine.
}
\examples{
f <- function(x) sqrt(x)
borne_inf <- -10
borne_sup <- 10
verifier_continuite(f, borne_inf, borne_sup)
}
